old-versions/requirements.txt
attrs==23.1.0
aws-cdk-lib==2.114.1
aws-cdk.asset-awscli-v1==2.2.201
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
blinker==1.7.0
cattrs==23.1.2
click==8.1.7
constructs==10.3.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsii==1.92.0
jsonify==0.5
MarkupSafe==2.1.3
marshmallow==3.20.1
mysql-connector-python==8.2.0
packaging==23.2
protobuf==4.21.12
publication==0.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pywatchman==1.4.1
six==1.16.0
SQLAlchemy==2.0.23
typeguard==2.13.3
typing_extensions==4.8.0
Werkzeug==3.0.1
